Remove hard-coded "mysql" string from metric names
The default scheme includes "mysql" so I do not think it needs to be hard-coded in the metric names.  Otherwise you end up with "#{Socket.gethostname}.mysql.mysql.category.metric".  Also, the slave metric does not have "mysql" hardcoded, so the default settings cause the slaveLag metric to be stored in a "general" directory as a peer to the "mysql" directory, instead of within the "mysql.general" directory with the rest of the metrics.