Added plugin which is checking AMQP connections
Some people don't use HTTP API of Rabbit or have that TCP port closed
so can't use it. This check is for clean AMQP connection however
depends on Bunny gem which is a pain. This was tested with the
following Bunny gem version: 0.9.0.pre9 which is basically
latest Bunny version at the time of creation of this plugin