Add ssl and basic auth to graphite/check-data.rb
We have added the ability to use ssl to communicate with graphite using
the --ssl boolean cli flag. If you are using self signed certs and need
to remove cert checking use --ssl-ignore-certs.

If you provide a username(-u, --username) and a password(-p, --password)
then basic auth will be used when connecting to graphite.