Add flag to ignore no data errors from graphite
Sometime we do not care if graphite does not return any data for a given
time period. For example if we are monitoring 5xx and there weren't any
then we do not want to critical.

I have added a --ignore-no-data flag to check-data.rb which will supress
the raised critical.