Events can explicitly set tags
It would be useful to be able to have a checks set an array of tags that should be associated with a check for use by flapjack.  For instance if you want influence what flapjack does with a certain check you need to have specific tags associated with it.

Including something the following in a sensu event payload
"tags"   => ['bypass_ncsm','party_animal'],

Will then be available to flapjack when it goes to process the event.