mailer.rb: Add string status to subject line
The subject line was set to contain @event['check']['notification'],
which does not exist in event data (in current versions).  Instead, put
a textual status line in there (like Nagios traditionally does), so it's
easier to distinguish warnings and critical from the subject line alone.
Also convert the status in the body of the message to a string.