Add process restart check
This sensu check is used to check if a running process requires a
restart if a library/package has been changed. This is useful you auto
upgrade pacakages and need to know which services need to be restarted
as a result of the upgrade. The default is to Warn if a single process
needs to be restarted and Critical if 2 or more processes need to be restarted.

The implementation of this check makes it Debian specific and will not work on
distributions/OS. It also requires passworless sudo for the Sensu user.