Ensure the metrics come in as an array
I had an issue where `@event['check']['output']` was a string value.  I don't know if this is always the case, or just for the `cpu_usage` metric I am trying out, but coercing it to be an array fixed the issue for me.  This should have no effect if the metrics are already an array.