Allow from field to be specified for clockworksms
Add support to the clockworksms handler for a "from" field which can be
specified to allow customisation of where the SMS appears to be sent
from.

Before this, it would use whatever the default "from" setting was set
on the account.  Now it can be specified in the json settings file, but
if not specified then it will default to "SENSU".