Do not create lock for GPG key check
In most cases the sensu client will be as a different user that owns the
GPG home directory and will not have write permissions to the home
directory. As we are only reading keys, we don't need (or want) to
create a lock file. This means we can give read permission to the GPG
homedir without giving write permission.

In future, it might be worth considering adding an option for the user
and using sudo to run the command.

Yasser