Setup office hours mutator to be handle time offsets
I plan to do additional work to 1) support DST and 2) load the
configuration of what you consider office hours from JSON.  This puts
the mutator in better shape now though.  Most servers are in GMT and the
out of the box setupi for this mutator is broken.  With this change all you have to do is change the gmt_offset variable to correctly align your office time vs. your servers time.  Currently this mutator would only work if you worked in GMT.  I also removed the Rubygems require since that's only needed for Ruby 1.8.7 and the removal of hash rockets in the plugins / handlers broke that support anyways.