Fix broken instance event check
The logic in https://github.com/sensu/sensu-community-plugins/commit/04106ed83ef769470512e3a1dd2d56e1b772f957 was flawed and it causes this check to never return any instances with events.  Instances are being excluded if the select finds a reboot completed, but an empty array satisfies the unless so it never finds any instances even if they events.  I've tested this change in logic in my environment to show the nodes Amazon is about to reboot.  This should get merged ASAP since the event is upcoming for everyone.