Update system_profile.rb
The extension works well, but it's missing a few things.

1. A general prefix, before the client name. This is typically aws.region.vpc-number.subnet. I can achieve this with path_prefix, and with setting prefix_at_start: 1

2. A way to contain these metrics in their own branch under client name. Again, path_prefix could be used, with prefix_at_start: 0, but it's taken already.

So I've added metric_prefix, which now goes below client name but above these metrics. I think it's a lot more flexible this way, and makes for a more organized namespace.

I've also implemented the disk_usage function, based on the sys-filesystem gem. It's lightweight, it doesn't spawn processes when executed, and captures the percentage of disk space used by files on each volume.

Tested and works fine with Sensu 0.16.0 and Graphite 0.9.13-pre1