Add a new plugin to monitor the time a RabbitMQ queue will take to drain
What's this drain time nonsense you speak of?  Monitoring queues is pretty tricky and traditionally people monitor pure message count.  Unfortunately this metric isn't very useful since 30,000 messages in a queue could be no big deal if you process 1,000 a second.  Rabbit exposes the rate of egress in queues.  From this it's pretty easy to determine the time a queue will take to drain to 0 if the current count and egress rate continue.  This allows for a more useful monitoring threshold where you alert on the time the queue will be backed up.